+++ app/pages/user/admin/[id].vue
@@ -43,7 +43,7 @@
 
                     <div class="mb-4">
                         <span class="font-medium text-900 mb-2">Profile Picture</span><br>
-                        <Avatar :image="user.meta?.avatarUrl" alt="user avatar" class="h-10rem w-10rem" shape="circle" />
+                        <Avatar :image="user.links?.avatarImage?.href" alt="user avatar" class="h-10rem w-10rem" shape="circle" />
                     </div>
                 </FieldsetItem>